aichatkit/requirements.txt
openai==1.12.0
google-generativeai==0.3.2
anthropic==0.25.1
